requirements.txt
fastapi
uvicorn
bcrypt==4.0.1
passlib
PyJWT
fastapi_controllers
pydantic[email]
python-dateutil